changes to delet_file
added file to delete files

+++ delete_file.php
@@ -4,13 +4,15 @@ ob_start();
 include_once('connection.php');
 include_once('other_funcs.php')
 
-$path = 'upload/'. $_FILES["file"]["name"];
+if($_POST['file_path']){
+	$path = $_POST['file_path'];
 
-if (is_file($path)){
-   if(unlink($path)){
-   	do_remove_file($_SESSION['email'], $path);
-   }else{
-   	echo "File could not be deleleted";
-   }
+	if (is_file($path)){
+	   if(unlink($path)){
+	   	do_remove_file($_SESSION['email'], $path);
+	   }else{
+	   	echo "File could not be deleleted";
+	   }
 
-  }
+	  }
+}
